Back_End/scraperapi/requirements.txt
Flask_Cors==3.0.10
Flask==1.1.2
selenium==3.141.0
requests==2.24.0
joblib==1.0.0
msedge_selenium_tools==3.141.3
beautifulsoup4==4.9.3

